Split CWL parser and converter
